COPL-609 - Numeric properties of estimators should be displayed in actual number fields COPL-486 - Implement dynamic property: numeric value as a spinner

+++ coko/com.elster.jupiter.estimators/src/test/java/com/elster/jupiter/estimators/impl/LinearInterpolationTest.java
@@ -104,7 +104,7 @@ public class LinearInterpolationTest {
     public void testLinearInterpolation() {
 
         Map<String, Object> properties = new HashMap<String, Object>();
-        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, BigDecimal.valueOf(10));
+        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, 10L);
 
         Estimator estimator = new LinearInterpolation(thesaurus, propertySpecService, properties);
         estimator.init(Logger.getAnonymousLogger());
@@ -123,7 +123,7 @@ public class LinearInterpolationTest {
     public void testLinearInterpolationDoesNotEstimateWhenTooManySuspects() {
 
         Map<String, Object> properties = new HashMap<String, Object>();
-        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, BigDecimal.valueOf(1));
+        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, 1L);
 
         Estimator estimator = new LinearInterpolation(thesaurus, propertySpecService, properties);
         estimator.init(Logger.getAnonymousLogger());
@@ -140,7 +140,7 @@ public class LinearInterpolationTest {
         doReturn(false).when(readingType).isCumulative();
 
         Map<String, Object> properties = new HashMap<String, Object>();
-        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, BigDecimal.valueOf(10));
+        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, 10L);
 
         Estimator estimator = new LinearInterpolation(thesaurus, propertySpecService, properties);
         estimator.init(Logger.getAnonymousLogger());
@@ -157,7 +157,7 @@ public class LinearInterpolationTest {
         doReturn(Optional.empty()).when(channel).getReading(BEFORE.toInstant());
 
         Map<String, Object> properties = new HashMap<String, Object>();
-        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, BigDecimal.valueOf(10));
+        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, 10L);
 
         Estimator estimator = new LinearInterpolation(thesaurus, propertySpecService, properties);
         estimator.init(Logger.getAnonymousLogger());
@@ -174,7 +174,7 @@ public class LinearInterpolationTest {
         doReturn(Optional.empty()).when(channel).getReading(AFTER.toInstant());
 
         Map<String, Object> properties = new HashMap<>();
-        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, BigDecimal.valueOf(10));
+        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, 10L);
 
         Estimator estimator = new LinearInterpolation(thesaurus, propertySpecService, properties);
         estimator.init(Logger.getAnonymousLogger());
@@ -191,7 +191,7 @@ public class LinearInterpolationTest {
         doReturn(null).when(readingRecord1).getQuantity(readingType);
 
         Map<String, Object> properties = new HashMap<>();
-        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, BigDecimal.valueOf(10));
+        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, 10L);
 
         Estimator estimator = new LinearInterpolation(thesaurus, propertySpecService, properties);
         estimator.init(Logger.getAnonymousLogger());
@@ -208,7 +208,7 @@ public class LinearInterpolationTest {
         doReturn(null).when(readingRecord2).getQuantity(readingType);
 
         Map<String, Object> properties = new HashMap<>();
-        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, BigDecimal.valueOf(10));
+        properties.put(MAX_NUMBER_OF_CONSECUTIVE_SUSPECTS, 10L);
 
         Estimator estimator = new LinearInterpolation(thesaurus, propertySpecService, properties);
         estimator.init(Logger.getAnonymousLogger());
